Refactor proxy pass process handling and logging

+++ hikka/web/proxypass.py
@@ -6,7 +6,6 @@
 
 import asyncio
 import logging
-import os
 import re
 import typing
 
@@ -33,13 +32,15 @@ class ProxyPasser:
         for getline in iter(stream.readline, ""):
             await asyncio.sleep(delay)
             data_chunk = await getline
+
             if await callback(data_chunk.decode("utf-8")):
                 if not self._url_available.is_set():
                     self._url_available.set()
 
     def kill(self):
         try:
-            self._sproc.terminate()
+            if self._sproc:
+                self._sproc.terminate()
         except Exception:
             logger.exception("Failed to kill proxy pass process")
         else:
@@ -47,40 +48,57 @@ class ProxyPasser:
 
     async def _process_stream(self, stdout_line: str) -> None:
         logger.debug(stdout_line)
-        regex = r"https://[^\s]+"
 
-        if re.search(regex, stdout_line):
-            self._tunnel_url = re.search(regex, stdout_line)[0].rstrip("./,;)")
+        regex = r"https://[^\s]+"
+        match = re.search(regex, stdout_line)
+
+        if match:
+            self._tunnel_url = match.group(0).rstrip("./,;)")
             self._change_url_callback(self._tunnel_url)
-            logger.debug("Proxy pass tunneled: %s", self._tunnel_url)
+
+            logger.debug(
+                "Proxy pass tunneled: %s",
+                self._tunnel_url,
+            )
+
             self._url_available.set()
 
-    async def get_url(self, port: int, no_retry: bool = False) -> typing.Optional[str]:
+    async def get_url(
+        self,
+        port: int,
+        no_retry: bool = False,
+    ) -> typing.Optional[str]:
         async with self._lock:
-            if self._tunnel_url:
+
+            if self._tunnel_url and self._sproc:
                 try:
-                    await asyncio.wait_for(self._sproc.wait(), timeout=0.05)
+                    await asyncio.wait_for(
+                        self._sproc.wait(),
+                        timeout=0.05,
+                    )
+
                 except asyncio.TimeoutError:
                     return self._tunnel_url
+
                 else:
                     self.kill()
+                    self._tunnel_url = None
 
-            if "DOCKER" in os.environ:
-                # We're in a Docker container, so we can't use ssh
-                # Also, the concept of Docker is to keep
-                # everything isolated, so we can't proxy-pass to
-                # open web.
-                return None
+            logger.debug(
+                "Starting proxy pass shell for port %d",
+                port,
+            )
 
-            logger.debug("Starting proxy pass shell for port %d", port)
             self._sproc = await asyncio.create_subprocess_shell(
                 (
-                    "ssh -o StrictHostKeyChecking=no "
+                    "ssh "
+                    "-o StrictHostKeyChecking=no "
                     "-o UserKnownHostsFile=/dev/null "
                     "-o ExitOnForwardFailure=yes "
                     "-o ServerAliveInterval=30 "
-                    "-R"
-                    f" 80:127.0.0.1:{port} nokey@localhost.run"
+                    "-R "
+                    f"80:127.0.0.1:{port} "
+                    "nokey@localhost.run"
                 ),
                 stdin=asyncio.subprocess.PIPE,
                 stdout=asyncio.subprocess.PIPE,
@@ -90,8 +108,16 @@ class ProxyPasser:
             utils.atexit(self.kill)
 
             self._url_available = asyncio.Event()
-            logger.debug("Starting proxy pass reader for port %d", port)
-            for stream in (self._sproc.stdout, self._sproc.stderr):
+
+            logger.debug(
+                "Starting proxy pass reader for port %d",
+                port,
+            )
+
+            for stream in (
+                self._sproc.stdout,
+                self._sproc.stderr,
+            ):
                 asyncio.ensure_future(
                     self._read_stream(
                         self._process_stream,
@@ -101,11 +127,19 @@ class ProxyPasser:
                 )
 
             try:
-                await asyncio.wait_for(self._url_available.wait(), 20)
+                await asyncio.wait_for(
+                    self._url_available.wait(),
+                    20,
+                )
+
             except asyncio.TimeoutError:
                 self.kill()
                 self._tunnel_url = None
 
-            logger.debug("Proxy pass tunnel url to port %d: %s", port, self._tunnel_url)
+            logger.debug(
+                "Proxy pass tunnel url to port %d: %s",
+                port,
+                self._tunnel_url,
+            )
 
             return self._tunnel_url
